Move in-memory event store and tracker into a Memory namespace
This matches the Postgres version.
