+++ lib/event_sourcery.rb
@@ -8,7 +8,6 @@ require 'event_sourcery/event_store/event_sink'
 require 'event_sourcery/event_store/event_source'
 require 'event_sourcery/errors'
 require 'event_sourcery/event_store/each_by_range'
-require 'event_sourcery/event_store/memory'
 require 'event_sourcery/event_store/subscription'
 require 'event_sourcery/event_store/poll_waiter'
 require 'event_sourcery/event_store/event_builder'
@@ -22,13 +21,14 @@ require 'event_sourcery/event_processing/error_handlers/constant_retry'
 require 'event_sourcery/event_processing/error_handlers/exponential_backoff_retry'
 require 'event_sourcery/event_processing/esp_process'
 require 'event_sourcery/event_processing/esp_runner'
-require 'event_sourcery/event_processing/event_trackers/memory'
 require 'event_sourcery/event_processing/event_stream_processor'
 require 'event_sourcery/event_processing/event_stream_processor_registry'
 require 'event_sourcery/config'
 require 'event_sourcery/event_body_serializer'
 require 'event_sourcery/aggregate_root'
 require 'event_sourcery/repository'
+require 'event_sourcery/memory/tracker'
+require 'event_sourcery/memory/event_store'
 
 module EventSourcery
   def self.configure
